Rename int to i64

+++ src/analyzer/program.rs
@@ -72,7 +72,7 @@ mod tests {
     fn variable_assignment() {
         let raw = r#"
         fn main() {
-            int i = 10
+            i64 i = 10
             i = 11
         }
         "#;
@@ -86,7 +86,7 @@ mod tests {
     fn fn_decl() {
         let raw = r#"
         fn main() {}
-        fn test(int a, string b) { 
+        fn test(i64 a, string b) { 
             string s = "hello world!" 
         }"#;
         let mut tokens = Token::tokenize(Cursor::new(raw).lines()).expect("should not error");
@@ -137,12 +137,12 @@ mod tests {
     fn big_program() {
         let raw = r#"
         fn main() {
-            int i = 0
+            i64 i = 0
             loop {
                 string prefix = str_concat(str_concat("Fibonacci number ", itoa(i)), " is: ")
-                int result = fib(
+                i64 result = fib(
                     i,
-                    fn (int n): bool {
+                    fn (i64 n): bool {
                         print(str_concat("fib visitor sees n=", itoa(n)))
                         return n % 2 == 0
                     },
@@ -155,7 +155,7 @@ mod tests {
         }
         
         // Calls `visitor_fn` with n and returns the n'th Fibonacci number.
-        fn fib(int n, fn (int): bool visitor_fn): int {
+        fn fib(i64 n, fn (i64): bool visitor_fn): i64 {
             if visitor_fn(n) {
                 print("visitor returned true")
             }
@@ -171,7 +171,7 @@ mod tests {
             return a
         }
         
-        fn itoa(int i): string {
+        fn itoa(i64 i): string {
             return "fake"
         }
         "#;
